shell structure changes and branch rename
